# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Gestion Magasin Documentation",
        "description": "Implementation solution de Gestion de Magasin avec Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/devis": {
            "get": {
                "tags": [
                    "Affiche devis et lignedevis"
                ],
                "summary": "Affiche la liste des devis et lignedevis",
                "description": "Retourner Touts les devis et lignedevis",
                "operationId": "Affiche devis et lignedevis",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ajouter devis et lignedevis"
                ],
                "summary": "Ajoute devis et lignedevis",
                "description": "Ajouter  devis et lignedevis et Retourne les Valeurs Ajouter",
                "operationId": "Store devis et lignedevis",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/devis/{devis}": {
            "get": {
                "tags": [
                    "Affiche devis et lignedevis par ID"
                ],
                "summary": "Affiche le devis et lignedevis par ID",
                "description": "Retourner le devis et lignedevis",
                "operationId": "Show devis et lignedevis",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mise en Jour devis et lignedevis"
                ],
                "summary": "Modifier devis et lignedevis",
                "description": "Retourner devis et lignedevis Modifier",
                "operationId": "Modifier devis et lignedevis",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supprimer devis et lignedevis"
                ],
                "summary": "Supprimer un devis et lignedevis",
                "description": "Supprimer un devis et lignedevis",
                "operationId": "Delete devis et lignedevis",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}